Implement memory cleanup with WeakCells
Implement a proof-of-concept fix for the memory leak using WeakFactory as it is currently implemented in Chrome Canary (73.0.3635.0). Note this API appears likely to change, but it works - GC of Proxy objects can be detected and messages sent to the receiver to delete entries from the ID map.

Also update terminology to "receiver" and "controller" rather than "main thread" and "worker", since it's intended to work either way round (worker controlling DOM or DOM controlling worker). Finally, remove the now-unnecessary ResetReferences method.